Oprava hostname, omezení na VirtualBox v /etc/dhcp/class.conf stále trvá

+++ masquarate.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 #testování vstupů
-[[ "$1" == ""  ]] && {
-	echo "Použití ./$0 <vnitřní eth rozhraní> <vnější eth rozhraní>"
+if [[ "$1" == ""  ]];then
+	echo "Použití ./$0 <vnitřní eth rozhraní> <vnější eth rozhraní>";
 	exit 3;
-}
+fi
 
 eths=(`ip link | awk 'BEGIN{FS=" "}{print $2}' | grep ".*:$" | tr -d ':'`);
 ff=0;
@@ -18,10 +18,10 @@ done;
 [[ "$sf" == "0" ]] && { exit 1; }
 
 [[ "$1" == "$2" ]] && { exit 2; }
-[[ -f "/etc/init.d/iptables-persistent" ]] || {
+if [[ -f "/etc/init.d/iptables-persistent" ]];then
 	export DEBIAN_FRONTEND=noninteractive
 	apt-get install iptables-persistent -y
-}
+fi
 # nastavení iptables
 /sbin/iptables -t nat -A POSTROUTING -o "$2" -j MASQUERADE
 /sbin/iptables -A FORWARD -i "$2" -o "$1" -m state --state RELATED,ESTABLISHED -j ACCEPT
@@ -32,16 +32,16 @@ tm=`mktemp`;
 ins=0;
 # nastavení routování napořád
 while read line;do
-	[[ $line  =~ net.ipv4.ip_forward ]] && {
+	if [[ $line  =~ net.ipv4.ip_forward ]];then
 		echo "net.ipv4.ip_forward = 1" >> "$tm";
 		ins=1;
-	} || {
+	else
 		echo "$line" >> "$tm";
-	}
+	fi
 done < "/etc/sysctl.conf";
-[[ "$ins" == "0" ]] && {
+if [[ "$ins" == "0" ]];then
 	echo "net.ipv4.ip_forward = 1" >> "$tm";
-}
+fi
 rm "/etc/sysctl.conf";
 mv "$tm" "/etc/sysctl.conf";
 # ukládání iptables
